requirements-dev.txt
# 开发依赖
-r requirements.txt

# 可选依赖（用于LLM模式）
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0